move plugins to own test file
these should be gone once we do #99 and these will move into each
respective gem test suite
